plugin: do not keep track of parent objects
This change comes with a very small change of breaking things -
at least informational lists of stored objects and the stopgap for
not deleting a VM template if it has linked copies - but none of those
that we can think of right now will actually lead to any performance
problems or data loss.

TBH, using StorPool's volume parent for that was wrong anyway.