Track volumes and snapshots by their "name" field
Do not use a volume's globalId field as returned by the API, since that may
not be the same as the one that the volume had at first, e.g. it may change
after a volume is reverted to a snapshot. Instead, add the "name" field to
our internal (encoded/decoded) structure for easier comparison, and
compare against a volume's name field as returned by the API.

For snapshots there is no problem in using the globalId field, though this
may also be changed in the future.