Remove patching of PVE for cloud-init on StorPool support
Using the `vm-$vm_id-cloudinit-sp-$globalId.raw` PVE disk name format is
easier to implement, but also requires modifying a regular expression in the
upstream PVE source code in order to pass the `-sp-$globalId` suffix.
We are trying to avoid that and in this implementation use the SP volume tags
to match the cloud-init image volume for the given VM. The new disk name format
is `vm-$vm_id-cloudinit.raw`.

To be able to remove the PVE patch, you need to remove and add anew the
StorPool-backed cloud-init drive for every VM that uses one, which should be
a seamless operation. This version of the plugin works seamlessly with both
disk name formats.