Fix timeout on volume_size_info
When a client calls the PVE API endpoint
`nodes/{node}/storage/{storage}/content/{volume}`, the plugin would make a
very slow SP API call (`VolumesGetStatus`) in order to report the space used
inside the volume. This can cause timeouts and break external backups, even
though the value is almost expected to not be useful according to the PVE API
docs.

With this fix, the plugin uses a much faster API call, and returns the volume
size as both the size and used space. While technically not correct, it works,
and saves the effort of implementing caching of the expensive API call across
the cluster.