Fix VM live migration failure after OOM kill (#5)
During migration, the migration process is killed due to OOM, and the host is unable to live migrate the VM.
The problem is that the migration is still running on the destination node after it's parent get OOM killed and in the meantime the migration is flagged as failed. In this case the VM property lock is deleted and activate_volume() sub wont be aware that this is a migration and will force detach the volume from the source node VM.
Fixed by killing the worker PID if it's parent is dead.